// src/app/api/models/shipping-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AddressDto } from '../models/address-dto';
import { ShippingStatusDto } from '../models/shipping-status-dto';
export interface ShippingDto {
  actualDeliveryDate?: string;
  carrier: string;
  createdAt?: string;
  currentStatus?: string;
  estimatedDeliveryDate?: string;
  id?: number;
  notes?: string;
  orderId: number;
  orderNumber?: string;
  shippedDate?: string;
  shippingAddress: AddressDto;
  shippingCost?: number;
  shippingMethod: string;
  specialInstructions?: string;
  statusHistory?: Array<ShippingStatusDto>;
  trackingNumber?: string;
  updatedAt?: string;
  weight?: number;
}