// src/app/api/models/create-shipping-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateShippingRequest {
  carrier: string;
  notes?: string;
  orderId: number;
  shippingAddressId: number;
  shippingCost?: number;
  shippingMethod: string;
  specialInstructions?: string;
  weight?: number;
}